added export/import components serialization

+++ libs/InteractiveToolkit-AppKit/opengl/appkit-gl-engine/include/appkit-gl-engine/Renderer/LightAndShadowManager.h
@@ -75,7 +75,7 @@ namespace AppKit
                 // writer.EndObject();
                 
             }
-            void Deserialize(rapidjson::Value &_value, std::unordered_map<intptr_t, std::shared_ptr<Transform>> &transform_map, std::unordered_map<intptr_t, std::shared_ptr<Component>> &component_map){
+            void Deserialize(rapidjson::Value &_value, std::unordered_map<uint64_t, std::shared_ptr<Transform>> &transform_map, std::unordered_map<uint64_t, std::shared_ptr<Component>> &component_map){
                 // if (!_value.HasMember("type") || !_value["type"].IsString())
                 //     return;
                 // if (!strcmp(_value["type"].GetString(), ShadowCache::Type) == 0)
