+++ libs/InteractiveToolkit-AppKit/opengl/appkit-gl-engine/src/Components/ComponentFrustumCulling.cpp
@@ -306,7 +306,7 @@ namespace AppKit
                 writer.EndObject();
                 
             }
-            void ComponentFrustumCulling::Deserialize(rapidjson::Value &_value, std::unordered_map<intptr_t, std::shared_ptr<Transform>> &transform_map, std::unordered_map<intptr_t, std::shared_ptr<Component>> &component_map){
+            void ComponentFrustumCulling::Deserialize(rapidjson::Value &_value, std::unordered_map<uint64_t, std::shared_ptr<Transform>> &transform_map, std::unordered_map<uint64_t, std::shared_ptr<Component>> &component_map){
                 if (!_value.HasMember("type") || !_value["type"].IsString())
                     return;
                 if (!strcmp(_value["type"].GetString(), ComponentFrustumCulling::Type) == 0)
